Consolidate FilterSet classes
